feat(subscriber): refactor create subscriber form and integrate telemetry tracking
